Create mongo-secret.yml
apiVersion : this is the version of the API used by the cluster.

kind: kind define whether it’s a deployment, secret or configMap.

metadata: here we can define data about the object we are about to create. In this example, we only provide the name of the pod. But you can provide other details like the namespace.

type:  type Opaque means that from kubernetes’s point of view the contents of this Secret is unstructured, it can contain arbitrary key-value pairs.

data : data contains the data, you want to save as secrets.

Deploy the secret config file via kubectl apply -f mongo-secret.yaml